requirements.txt
pdfplumber
openai
numpy
python-dotenv
streamlit
